rfctr: extract matrix modules
No code changes other than renaming classes to make them public where
they now need to be accessed across module borders.
