+++ src/cr/cube/matrix/cubemeasure.py
@@ -1,872 +1,21 @@
 # encoding: utf-8
 
-"""A matrix is the 2D cube-data partition used by a slice.
+"""Provides abstracted cube-measure objects used as the basis for second-order measures.
 
-The `Assembler` object provides the external interface for this module. Its name derives
-from its role to "assemble" a finished 2D array ("matrix") for a particular measure from
-the base measure values and inserted subtotals, to reorder the rows and columns
-according to the dimension *order* transforms, and to hide rows and columns that are
-either hidden by the user or "pruned" because they contain no observations.
+There are several cube-measures that can appear in a cube-response, including
+unweighted-counts, weighted-counts (aka. counts), means, and others.
 """
 
 from __future__ import division
 
 import numpy as np
-from scipy.stats import norm
 from scipy.stats.contingency import expected_freq
 
-from cr.cube.collator import ExplicitOrderCollator, PayloadOrderCollator
-from cr.cube.enums import COLLATION_METHOD as CM, DIMENSION_TYPE as DT
+from cr.cube.enums import DIMENSION_TYPE as DT
 from cr.cube.util import lazyproperty
 
 
-class Assembler(object):
-    """Provides measure and margin methods for a cube-slice.
-
-    An assembled matrix is a 2D ndarray reflecting all ordering, insertion, and hiding
-    transforms applied to the dimensions. An assembled margin is often a 1D ndarray
-    which is similarly formed from inserted values, ordered, and value hiding applied.
-
-    `cube` is the `cr.cube.Cube` object containing the data for this matrix. Note that
-    not all the data in `cube` will necessarily be used by this matrix. When `cube` is
-    more than 2-dimensional, it is "sliced" and each slice gets its own matrix (and
-    `_Slice` object).
-
-    `dimensions` is a pair (2-tuple) of (rows_dimension, columns_dimension) Dimension
-    objects. These are always the last two dimensions of `cube` but may and often do
-    have transformations applied that are not present on the `cube` dimensions from
-    which they derive.
-
-    `slice_idx` is an int offset indicating which portion of `cube` data to use for this
-    matrix. There is one slice for each element of the first cube dimension (the "table"
-    dimension) when the cube has more than two dimensions.
-    """
-
-    def __init__(self, cube, dimensions, slice_idx):
-        self._cube = cube
-        self._dimensions = dimensions
-        self._slice_idx = slice_idx
-
-    @lazyproperty
-    def column_index(self):
-        """2D np.float64 ndarray of column-index "percentage" for each table cell."""
-        return self._assemble_matrix(
-            _NanSubtotals.blocks(self._cube_result_matrix, "column_index")
-        )
-
-    @lazyproperty
-    def column_labels(self):
-        """1D str ndarray of column name for each matrix column.
-
-        These are suitable for use as column headings; labels for subtotal columns
-        appear in the sequence and labels are ordered to correspond with their
-        respective data column.
-        """
-        return self._dimension_labels(self._columns_dimension, self._column_order)
-
-    @lazyproperty
-    def columns_base(self):
-        """1D/2D np.int64 ndarray of unweighted-N for each slice column/cell."""
-        # --- an MR_X slice produces a 2D column-base (each cell has its own N) ---
-        if self._rows_dimension.dimension_type == DT.MR_SUBVAR:
-            return self._assemble_matrix(
-                _SumSubtotals.blocks(self._cube_result_matrix, "columns_base")
-            )
-
-        # --- otherwise columns-base is a vector ---
-        return self._assemble_vector(
-            self._cube_result_matrix.columns_base,
-            self._column_subtotals,
-            self._column_order,
-        )
-
-    @lazyproperty
-    def columns_dimension_numeric_values(self):
-        """1D optional np.int/float64 ndarray of numeric-value for each column element.
-
-        A value of np.nan appears for a column element without a numeric-value. All
-        subtotal columns have a value of np.nan (subtotals have no numeric value).
-        """
-        elements = self._columns_dimension.valid_elements
-        return np.array(
-            [
-                (elements[idx].numeric_value if idx >= 0 else np.nan)
-                for idx in self._column_order
-            ]
-        )
-
-    @lazyproperty
-    def columns_margin(self):
-        """1D/2D np.float64 ndarray of weighted-N for each column of this slice."""
-        # --- an MR_X slice produces a 2D columns-margin (each cell has its own N) ---
-        if self._rows_dimension.dimension_type == DT.MR_SUBVAR:
-            return self._assemble_matrix(
-                _SumSubtotals.blocks(self._cube_result_matrix, "columns_margin")
-            )
-
-        # --- otherwise columns-base is a vector ---
-        return self._assemble_vector(
-            self._cube_result_matrix.columns_margin,
-            self._column_subtotals,
-            self._column_order,
-        )
-
-    @lazyproperty
-    def inserted_column_idxs(self):
-        """tuple of int index of each subtotal column in slice."""
-        return tuple(i for i, col_idx in enumerate(self._column_order) if col_idx < 0)
-
-    @lazyproperty
-    def inserted_row_idxs(self):
-        """tuple of int index of each subtotal row in slice."""
-        return tuple(i for i, row_idx in enumerate(self._row_order) if row_idx < 0)
-
-    @lazyproperty
-    def means(self):
-        """2D optional np.float64 ndarray of mean for each cell.
-
-        Raises `ValueError` if the cube-result does not include a means cube-measure.
-        """
-        if not self._cube.has_means:
-            raise ValueError("cube-result does not include a means cube-measure")
-        return self._assemble_matrix(
-            _NanSubtotals.blocks(self._cube_result_matrix, "means")
-        )
-
-    @lazyproperty
-    def pvalues(self):
-        """2D np.float64/np.nan ndarray of p-value for each matrix cell."""
-        return 2 * (1 - norm.cdf(np.abs(self.zscores)))
-
-    @lazyproperty
-    def row_labels(self):
-        """1D str ndarray of row name for each matrix row.
-
-        These are suitable for use as row headings; labels for subtotal rows appear in
-        the sequence and labels are ordered to correspond with their respective data
-        row.
-        """
-        return self._dimension_labels(self._rows_dimension, self._row_order)
-
-    @lazyproperty
-    def rows_base(self):
-        """1D/2D np.int64 ndarray of unweighted-N for each slice row/cell."""
-        # --- an X_MR slice produces a 2D row-base (each cell has its own N) ---
-        if self._columns_dimension.dimension_type == DT.MR_SUBVAR:
-            return self._assemble_matrix(
-                _SumSubtotals.blocks(self._cube_result_matrix, "rows_base")
-            )
-
-        # --- otherwise rows-base is a vector ---
-        return self._assemble_vector(
-            self._cube_result_matrix.rows_base, self._row_subtotals, self._row_order
-        )
-
-    @lazyproperty
-    def rows_dimension_fills(self):
-        """tuple of RGB str like "#def032" fill color for each row in slice."""
-        elements = self._rows_dimension.valid_elements
-        return tuple(
-            (elements[idx].fill if idx >= 0 else None) for idx in self._row_order
-        )
-
-    @lazyproperty
-    def rows_dimension_numeric_values(self):
-        """1D optional np.int/float64 ndarray of numeric-value for each row element.
-
-        A value of np.nan appears for a row element without a numeric-value. All
-        subtotal rows have a value of np.nan (subtotals have no numeric value).
-        """
-        elements = self._rows_dimension.valid_elements
-        return np.array(
-            [
-                (elements[idx].numeric_value if idx >= 0 else np.nan)
-                for idx in self._row_order
-            ]
-        )
-
-    @lazyproperty
-    def rows_margin(self):
-        """1D/2D np.float64 ndarray of weighted-N for each column of this slice.
-
-        This value is a 2D array if one or both of the slice dimensions are MR. This is
-        because each cell has a distinct margin in such a slice.
-        """
-        # --- an X_MR slice produces a 2D rows-margin (each cell has its own N) ---
-        if self._columns_dimension.dimension_type == DT.MR_SUBVAR:
-            return self._assemble_matrix(
-                _SumSubtotals.blocks(self._cube_result_matrix, "rows_margin")
-            )
-
-        # --- otherwise rows-margin is a vector ---
-        return self._assemble_vector(
-            self._cube_result_matrix.rows_margin, self._row_subtotals, self._row_order
-        )
-
-    @lazyproperty
-    def table_base(self):
-        """Scalar, 1D, or 2D ndarray of np.int64 unweighted-N for this slice.
-
-        This value has four distinct forms, depending on the slice dimensions:
-
-            * MR_X_MR - 2D ndarray with a distinct table-base value per cell.
-            * MR_X - 1D ndarray of value per *row* when only rows dimension is MR.
-            * X_MR - 1D ndarray of value per *column* when only columns dimension is MR.
-            * CAT_X_CAT - scalar float value when slice has no MR dimension.
-
-        """
-        # --- an MR dimension cannot have subtotals, and a non-scalar table base only
-        # --- arises in an MR dimension case, so a non-scalar table-base never has
-        # --- subtotals and we need only operate on the base values. There is no need to
-        # --- assemble with subtotals, however the values of a non-scalar table-base do
-        # --- still need to be ordered (which includes hiding).
-        base_table_base = self._cube_result_matrix.table_base
-
-        # --- MR_X_MR slice produces a 2D table-base (each cell has its own N) ---
-        if (
-            self._rows_dimension.dimension_type == DT.MR_SUBVAR
-            and self._columns_dimension.dimension_type == DT.MR_SUBVAR
-        ):
-            return base_table_base[np.ix_(self._row_order, self._column_order)]
-
-        # --- CAT_X_MR slice produces a 1D table-base (each column has its own N) ---
-        if self._columns_dimension.dimension_type == DT.MR_SUBVAR:
-            return base_table_base[self._column_order]
-
-        # --- MR_X_CAT slice produces a 1D table-base (each row has its own N) ---
-        if self._rows_dimension.dimension_type == DT.MR_SUBVAR:
-            return base_table_base[self._row_order]
-
-        # --- CAT_X_CAT produces scalar table-base ---
-        table_base = self._cube_result_matrix.table_base
-        return table_base
-
-    @lazyproperty
-    def table_base_unpruned(self):
-        """np.int64 scalar or a 1D or 2D ndarray of np.int64 representing table base.
-
-        This value includes hidden vectors, those with either a hide transform on
-        their element or that have been pruned (because their base (N) is zero). Also,
-        it does *not* include inserted subtotals. This does not affect a scalar value
-        but when the return value is an ndarray, the shape may be different than the
-        array returned by `.table_base`.
-
-        A multiple-response (MR) dimension produces an array of table-base values
-        because each element (subvariable) of the dimension represents a logically
-        distinct question which may not have been asked of all respondents. When both
-        dimensions are MR, the return value is a 2D ndarray. A CAT_X_CAT matrix produces
-        a scalar value for this property.
-        """
-        # TODO: This name is misleading. It's not only "unpruned" it's "before_hiding"
-        # (of either kind, prune or hide-transform). But the real problem is having this
-        # interface property at all. The need for this is related to expressing ranges
-        # for base and margin in cubes that have an MR dimension. The real solution is
-        # to compute ranges in `cr.cube` rather than leaking this sort of internal
-        # detail through the interface and making the client compute those for
-        # themselves. So this will require reconstructing that "show-ranges" requirement
-        # and either adding some sort of a `.range` property that returns a sequence of
-        # (min, max) tuples, or maybe just returning margin or base as tuples when
-        # appropriate and having something like a `.margin_is_ranges` predicate the
-        # client can switch on to control their rendering.
-        return self._cube_result_matrix.table_base
-
-    @lazyproperty
-    def table_margin(self):
-        """Scalar, 1D, or 2D ndarray of np.float64 weighted-N for this slice.
-
-        This value has four distinct forms, depending on the slice dimensions:
-
-            * MR_X_MR - 2D ndarray with a distinct table-margin value per cell.
-            * MR_X - 1D ndarray of value per *row* when only rows dimension is MR.
-            * X_MR - 1D ndarray of value per *column* when only columns dimension is MR.
-            * CAT_X_CAT - scalar float value when slice has no MR dimension.
-
-        """
-        # NOTE: This all works out even though it doesn't include subtotals and isn't
-        # assembled, because an MR dimension can't have subtotals. Because the
-        # dimensionality of table-margin "follows" the MR-dimension(s), it follows that
-        # a table-margin vector or matrix never has subtotals.
-
-        base_table_margin = self._cube_result_matrix.table_margin
-
-        # --- MR_X_MR slice produces a 2D table-margin (each cell has its own N) ---
-        if (
-            self._rows_dimension.dimension_type == DT.MR_SUBVAR
-            and self._columns_dimension.dimension_type == DT.MR_SUBVAR
-        ):
-            return base_table_margin[np.ix_(self._row_order, self._column_order)]
-
-        # --- MR_X_CAT slice produces a 1D table-margin (each row has its own N) ---
-        if self._rows_dimension.dimension_type == DT.MR_SUBVAR:
-            return base_table_margin[self._row_order]
-
-        # --- CAT_X_MR slice produces a 1D table-margin (each column has its own N) ---
-        if self._columns_dimension.dimension_type == DT.MR_SUBVAR:
-            return base_table_margin[self._column_order]
-
-        # --- CAT_X_CAT produces scalar table-margin ---
-        return base_table_margin
-
-    @lazyproperty
-    def table_margin_unpruned(self):
-        """np.float/int64 scalar or a 1D or 2D ndarray of np.float/int64 table margin.
-
-        This value includes hidden vectors, those with either a hide transform on
-        their element or that have been pruned (because their base (N) is zero). Also,
-        it does not include inserted subtotals. This
-        does not affect a scalar value but when the return value is an ndarray, the
-        shape may be different than the array returned by `.table_margin`.
-
-        A matrix with a multiple-response (MR) dimension produces an array of
-        table-margin values because each element (subvariable) of the dimension
-        represents a logically distinct question which may not have been asked of all
-        respondents. When both dimensions are MR, the return value is a 2D ndarray.
-        A CAT_X_CAT matrix produces a scalar value for this property.
-        """
-        # TODO: see TODO in `.table_base_unpruned`
-        return self._cube_result_matrix.table_margin
-
-    @lazyproperty
-    def table_stderrs(self):
-        """2D np.float64 ndarray of std-error of table-percent for each matrix cell."""
-        return self._assemble_matrix(
-            _TableStdErrSubtotals.blocks(self._cube_result_matrix)
-        )
-
-    @lazyproperty
-    def unweighted_counts(self):
-        """2D np.int64 ndarray of unweighted-count for each cell."""
-        return self._assemble_matrix(
-            _SumSubtotals.blocks(self._cube_result_matrix, "unweighted_counts")
-        )
-
-    @lazyproperty
-    def weighted_counts(self):
-        """2D np.float/int64 ndarray of weighted-count for each cell."""
-        return self._assemble_matrix(
-            _SumSubtotals.blocks(self._cube_result_matrix, "weighted_counts")
-        )
-
-    @lazyproperty
-    def zscores(self):
-        """2D np.float64 ndarray of std-res value for each cell of matrix.
-
-        A z-score is also known as a *standard score* and is the number of standard
-        deviations above (positive) or below (negative) the population mean a cell's
-        value is.
-        """
-        # --- punt except on CAT_X_CAT (if any MR dimensions it gets np.nans) ---
-        if (
-            self._rows_dimension.dimension_type == DT.MR_SUBVAR
-            or self._columns_dimension.dimension_type == DT.MR_SUBVAR
-        ):
-            return self._assemble_matrix(
-                _NanSubtotals.blocks(self._cube_result_matrix, "zscores")
-            )
-
-        return self._assemble_matrix(_ZscoreSubtotals.blocks(self._cube_result_matrix))
-
-    def _assemble_matrix(self, blocks):
-        """Return 2D ndarray matrix assembled from `blocks`.
-
-        The assembled matrix includes inserted vectors (rows and columns), has hidden
-        vectors removed, and is ordered by whatever sort method is applied in the
-        dimension transforms.
-        """
-        # --- These are assembled into a single 2D array, and then rearranged based on
-        # --- row and column orders. All insertion, ordering, and hiding transforms are
-        # --- reflected in the row and column orders. They each include (negative)
-        # --- insertion idxs, hidden and pruned vector indices have been removed, and
-        # --- the ordering method has been applied to determine the sequence each idx
-        # --- appears in. This directly produces a final array that is exactly the
-        # --- desired output.
-        return np.block(blocks)[np.ix_(self._row_order, self._column_order)]
-
-    def _assemble_vector(self, base_vector, subtotals, order):
-        """Return 1D ndarray of `base_vector` with inserted `subtotals`, in `order`.
-
-        Each subtotal value is the result of applying np.sum to the addends extracted
-        from `base_vector` according the the `addend_idxs` property of each subtotal in
-        `subtotals`. The returned array is arranged by `order`, including possibly
-        removing hidden or pruned values.
-        """
-        # TODO: This only works for "sum" subtotals, which is all that it needs so far,
-        # but a fuller solution will probably get the subtotal values from a
-        # _BaseSubtotals subclass.
-        vector_subtotals = np.array(
-            [np.sum(base_vector[subtotal.addend_idxs]) for subtotal in subtotals]
-        )
-        return np.hstack([base_vector, vector_subtotals])[order]
-
-    @lazyproperty
-    def _column_order(self):
-        """1D np.int64 ndarray of signed int idx for each assembled column.
-
-        Negative values represent inserted subtotal-column locations.
-        """
-        order = self._dimension_order(self._columns_dimension, self._empty_column_idxs)
-        return order[order >= 0] if self._prune_subtotal_columns else order
-
-    @lazyproperty
-    def _column_subtotals(self):
-        """Sequence of _Subtotal object for each inserted column."""
-        return self._columns_dimension.subtotals
-
-    @lazyproperty
-    def _columns_dimension(self):
-        """The `Dimension` object representing column elements in this matrix."""
-        return self._dimensions[1]
-
-    @lazyproperty
-    def _cube_result_matrix(self):
-        """_BaseCubeResultMatrix subclass object appropriate to this cube-slice.
-
-        This matrix object encapsulates cube-result array parsing and MR multi-value
-        differences and provides a foundational set of second-order analysis measure and
-        margin arrays.
-        """
-        return _BaseCubeResultMatrix.factory(
-            self._cube, self._dimensions, self._slice_idx
-        )
-
-    def _dimension_labels(self, dimension, order):
-        """1D str ndarray of name for each vector of `dimension`.
-
-        These are suitable for use as row/column headings; labels for subtotals appear
-        in the sequence specified by `order`, such that labels are ordered to correspond
-        with their respective data vector.
-        """
-        return np.array(
-            [e.label for e in dimension.valid_elements]
-            + [s.label for s in dimension.subtotals]
-        )[order]
-
-    def _dimension_order(self, dimension, empty_idxs):
-        """1D np.int64 ndarray of signed int idx for each vector in `dimension`.
-
-        Negative values represent inserted-vector locations. Returned sequence reflects
-        insertion, hiding, pruning, and ordering transforms specified in `dimension`.
-        """
-        collation_method = dimension.collation_method
-
-        # --- this will become a significantly more sophisticated dispatch when
-        # --- sort-by-value is implemented in later commits. In particular, the
-        # --- signature for the sort-by-value collator differs from the payload and
-        # --- explicit-order collators. So we'll defer refactoring this until that
-        # --- collator is added. (sc)
-        display_order = (
-            ExplicitOrderCollator.display_order(dimension, empty_idxs)
-            if collation_method == CM.EXPLICIT_ORDER
-            else PayloadOrderCollator.display_order(dimension, empty_idxs)
-        )
-
-        # --- Returning as np.array suits its intended purpose, which is to participate
-        # --- in an np._ix() call. It works fine as a sequence too for any alternate
-        # --- use. Specifying int type prevents failure when there are zero elements.
-        return np.array(display_order, dtype=int)
-
-    @lazyproperty
-    def _empty_column_idxs(self):
-        """tuple of int index for each column with (unweighted) N = 0.
-
-        These columns are subject to pruning, depending on a user setting in the
-        dimension.
-        """
-        pruning_base = self._cube_result_matrix.columns_pruning_base
-        return tuple(i for i, N in enumerate(pruning_base) if N == 0)
-
-    @lazyproperty
-    def _empty_row_idxs(self):
-        """tuple of int index for each row with N = 0.
-
-        These rows are subject to pruning, depending on a user setting in the dimension.
-        """
-        pruning_base = self._cube_result_matrix.rows_pruning_base
-        return tuple(i for i, N in enumerate(pruning_base) if N == 0)
-
-    @lazyproperty
-    def _prune_subtotal_columns(self):
-        """True if subtotal columns need to be pruned, False otherwise.
-
-        Subtotal-columns need to be pruned when all base-rows are pruned.
-        """
-        if not self._rows_dimension.prune:
-            return False
-
-        return len(self._empty_row_idxs) == len(self._rows_dimension.element_ids)
-
-    @lazyproperty
-    def _prune_subtotal_rows(self):
-        """True if subtotal rows need to be pruned, False otherwise.
-
-        Subtotal-rows need to be pruned when all base-columns are pruned.
-        """
-        if not self._columns_dimension.prune:
-            return False
-
-        return len(self._empty_column_idxs) == len(self._columns_dimension.element_ids)
-
-    @lazyproperty
-    def _row_order(self):
-        """1D np.int64 ndarray of signed int idx for each assembled row.
-
-        Negative values represent inserted subtotal-row locations.
-        """
-        order = self._dimension_order(self._rows_dimension, self._empty_row_idxs)
-        if self._prune_subtotal_rows:
-            return np.array([idx for idx in order if idx >= 0], dtype=int)
-        return order
-
-    @lazyproperty
-    def _row_subtotals(self):
-        """Sequence of _Subtotal object for each inserted row."""
-        return self._rows_dimension.subtotals
-
-    @lazyproperty
-    def _rows_dimension(self):
-        """The `Dimension` object representing row elements in this matrix."""
-        return self._dimensions[0]
-
-
-# === SUBTOTALS OBJECTS ===
-
-
-class _BaseSubtotals(object):
-    """Base class for Subtotals objects."""
-
-    def __init__(self, cube_result_matrix, measure_propname):
-        self._cube_result_matrix = cube_result_matrix
-        self._measure_propname = measure_propname
-
-    @classmethod
-    def blocks(cls, cube_result_matrix, measure_propname=None):
-        """Return base, row and col insertion, and intersection matrices.
-
-        These are in the form ready for assembly.
-        """
-        return cls(cube_result_matrix, measure_propname)._blocks
-
-    @lazyproperty
-    def _blocks(self):
-        """base, row and col insertion, and intersection matrices."""
-        return [
-            [self._base_values, self._subtotal_columns],
-            [self._subtotal_rows, self._intersections],
-        ]
-
-    @lazyproperty
-    def _base_counts(self):
-        """2D np.float64 ndarray of weighted-count for each cell of base matrix."""
-        return self._cube_result_matrix.weighted_counts
-
-    @lazyproperty
-    def _base_values(self):
-        """2D ndarray of "body" values from cube-result matrix."""
-        if self._measure_propname is None:
-            raise NotImplementedError(
-                "`%s` must implement `._base_values`" % type(self).__name__
-            )  # pragma: no cover
-        return getattr(self._cube_result_matrix, self._measure_propname)
-
-    @lazyproperty
-    def _column_subtotals(self):
-        """Sequence of _Subtotal object for each subtotal in columns-dimension."""
-        return self._cube_result_matrix.columns_dimension.subtotals
-
-    @lazyproperty
-    def _dtype(self):
-        """Numpy data-type for result matrices, used for empty arrays."""
-        return np.float64
-
-    def _intersection(self, row_subtotal, column_subtotal):
-        """Value for this row/column subtotal intersection."""
-        raise NotImplementedError(
-            "`%s` must implement `._intersection()`" % type(self).__name__
-        )  # pragma: no cover
-
-    @lazyproperty
-    def _intersections(self):
-        """(n_row_subtotals, n_col_subtotals) ndarray of intersection values.
-
-        An intersection value arises where a row-subtotal crosses a column-subtotal.
-        """
-        return np.array(
-            [
-                self._intersection(row_subtotal, column_subtotal)
-                for row_subtotal in self._row_subtotals
-                for column_subtotal in self._column_subtotals
-            ],
-            dtype=self._dtype,
-        ).reshape(len(self._row_subtotals), len(self._column_subtotals))
-
-    @lazyproperty
-    def _ncols(self):
-        """int count of columns in base-matrix."""
-        return self._base_values.shape[1]
-
-    @lazyproperty
-    def _nrows(self):
-        """int count of rows in base-matrix."""
-        return self._base_values.shape[0]
-
-    @lazyproperty
-    def _row_subtotals(self):
-        """Sequence of _Subtotal object for each subtotal in rows-dimension."""
-        return self._cube_result_matrix.rows_dimension.subtotals
-
-    def _subtotal_column(self, subtotal):
-        """Return (n_rows,) ndarray of values for `subtotal` column."""
-        raise NotImplementedError(
-            "`%s` must implement `._subtotal_column()`" % type(self).__name__
-        )  # pragma: no cover
-
-    @lazyproperty
-    def _subtotal_columns(self):
-        """(n_rows, n_col_subtotals) matrix of subtotal columns."""
-        subtotals = self._column_subtotals
-
-        if len(subtotals) == 0:
-            return np.empty((self._nrows, 0), dtype=self._dtype)
-
-        return np.hstack(
-            [
-                self._subtotal_column(subtotal).reshape(self._nrows, 1)
-                for subtotal in subtotals
-            ]
-        )
-
-    def _subtotal_row(self, subtotal):
-        """Return (n_cols,) ndarray of values for `subtotal` row."""
-        raise NotImplementedError(
-            "`%s` must implement `._subtotal_row()`" % type(self).__name__
-        )  # pragma: no cover
-
-    @lazyproperty
-    def _subtotal_rows(self):
-        """(n_row_subtotals, n_cols) ndarray of subtotal rows."""
-        subtotals = self._row_subtotals
-
-        if len(subtotals) == 0:
-            return np.empty((0, self._ncols), dtype=self._dtype)
-
-        return np.vstack(
-            [self._subtotal_row(subtotal) for subtotal in self._row_subtotals]
-        )
-
-    @lazyproperty
-    def _table_margin(self):
-        """Scalar or ndarray table-margin of cube-result matrix."""
-        return self._cube_result_matrix.table_margin
-
-
-class _NanSubtotals(_BaseSubtotals):
-    """Subtotal blocks for measures that cannot meaningfully be subtotaled.
-
-    Each subtotal value (and intersection value) is `np.nan`.
-    """
-
-    def _intersection(self, row_subtotal, column_subtotal):
-        """Unconditionally return np.nan for each intersection cell."""
-        return np.nan
-
-    def _subtotal_column(self, subtotal):
-        """Return (n_rows,) ndarray of np.nan values."""
-        return np.full(self._nrows, np.nan)
-
-    def _subtotal_row(self, subtotal):
-        """Return (n_cols,) ndarray of np.nan values."""
-        return np.full(self._ncols, np.nan)
-
-
-class _SumSubtotals(_BaseSubtotals):
-    """Subtotal "blocks" created by np.sum() on addends, primarily counts."""
-
-    @lazyproperty
-    def _dtype(self):
-        """Numpy data-type for result matrices, used for empty arrays."""
-        return self._base_values.dtype
-
-    def _intersection(self, row_subtotal, column_subtotal):
-        """Sum for this row/column subtotal intersection."""
-        return np.sum(self._subtotal_row(row_subtotal)[column_subtotal.addend_idxs])
-
-    def _subtotal_column(self, subtotal):
-        """Return (n_rows,) ndarray of values for `subtotal` column."""
-        return np.sum(self._base_values[:, subtotal.addend_idxs], axis=1)
-
-    def _subtotal_row(self, subtotal):
-        """Return (n_cols,) ndarray of values for `subtotal` row."""
-        return np.sum(self._base_values[subtotal.addend_idxs, :], axis=0)
-
-
-class _TableStdErrSubtotals(_BaseSubtotals):
-    """Computes subtotal values for the table-stderrs measure."""
-
-    @lazyproperty
-    def _base_values(self):
-        """2D np.float64 ndarray of table-stderr for each cell of cube-result matrix."""
-        return self._cube_result_matrix.table_stderrs
-
-    def _intersection(self, row_subtotal, column_subtotal):
-        """Return value for intersection of `row_subtotal` and `column_subtotal`."""
-        # --- column of base subtotal counts is 1D, like [44 148 283 72] ---
-        # --- the inserted-row for counts measure, like: [159 172 107 272] ---
-        row_subtotal_counts = np.sum(
-            self._base_counts[row_subtotal.addend_idxs, :], axis=0
-        )
-        intersection_count = np.sum(row_subtotal_counts[column_subtotal.addend_idxs])
-
-        # --- `p` is intersection-table-proportion ---
-        p = intersection_count / self._table_margin
-        table_proportion_variance = p * (1 - p)
-
-        return np.sqrt(table_proportion_variance / self._table_margin)
-
-    def _subtotal_column(self, subtotal):
-        """Return (n_rows,) ndarray of table-stderr `subtotal` value."""
-        # --- column of base subtotal counts is 1D, like [44 148 283 72] ---
-        subtotal_counts = np.sum(
-            self._base_counts[:, subtotal.addend_idxs],
-            axis=1,
-        )
-
-        # --- `p` is subtotal-table-proportions ---
-        p = subtotal_counts / self._table_margin
-        table_proportion_variance = p * (1 - p)
-
-        return np.sqrt(table_proportion_variance / self._table_margin)
-
-    def _subtotal_row(self, subtotal):
-        """Return (n_cols,) ndarray of table-stderr `subtotal` value."""
-        # --- row of base subtotal counts is 1D, like [435 392 260 162] ---
-        subtotal_counts = np.sum(self._base_counts[subtotal.addend_idxs, :], axis=0)
-
-        # --- `p` is subtotal-table-proportions ---
-        p = subtotal_counts / self._table_margin
-        table_proportion_variance = p * (1 - p)
-
-        return np.sqrt(table_proportion_variance / self._table_margin)
-
-
-class _ZscoreSubtotals(_BaseSubtotals):
-    """Computes subtotal values for the z-score measure.
-
-    This is only operative for a CAT_X_CAT cube-result; an MR dimension causes all
-    subtotals to be computed (elsewhere) as `np.nan`.
-    """
-
-    @lazyproperty
-    def _base_values(self):
-        """2D np.float64 ndarray of z-score for each cell of cube-result matrix."""
-        return self._cube_result_matrix.zscores
-
-    def _intersection(self, row_subtotal, column_subtotal):
-        """Return value for intersection of `row_subtotal` and `column_subtotal`."""
-        row_subtotal_counts = np.sum(
-            self._base_counts[row_subtotal.addend_idxs, :], axis=0
-        )
-
-        # --- the weighted-counts version of this intersection cell ---
-        intersection_count = np.sum(row_subtotal_counts[column_subtotal.addend_idxs])
-
-        # --- subtotal-margin is scalar, like 547 ---
-        subtotal_margin = np.sum(row_subtotal_counts)
-
-        # --- opposite-margin is scalar because no MR dimensions ---
-        opposite_margin = np.sum(
-            self._cube_result_matrix.columns_margin[column_subtotal.addend_idxs]
-        )
-
-        # --- table_margin is scalar because no MR dimensions ---
-        table_margin = self._table_margin
-
-        # --- expected_count is scalar ---
-        expected_count = opposite_margin * subtotal_margin / table_margin
-
-        # --- residuals is scalar ---
-        residuals = intersection_count - expected_count
-
-        # --- variance is scalar (I think) ---
-        variance = (
-            opposite_margin
-            * subtotal_margin
-            * ((table_margin - opposite_margin) * (table_margin - subtotal_margin))
-            / table_margin ** 3
-        )
-
-        # --- result is scalar ---
-        with np.errstate(divide="ignore", invalid="ignore"):
-            return residuals / np.sqrt(variance)
-
-    def _subtotal_column(self, subtotal):
-        """Return (n_rows,) ndarray of zscore `subtotal` value."""
-        # --- the weighted-counts version of this subtotal-column ---
-        subtotal_counts = np.sum(self._base_counts[:, subtotal.addend_idxs], axis=1)
-
-        # --- subtotal-margin is scalar, like 547 ---
-        subtotal_margin = np.sum(subtotal_counts)
-
-        # --- base-rows-margin is 1D because no MR dimensions ---
-        opposing_margin = self._cube_result_matrix.rows_margin
-
-        # --- table_margin is scalar because no MR dimensions ---
-        table_margin = self._table_margin
-
-        # --- expected_counts is 1D ---
-        expected_counts = opposing_margin * subtotal_margin / table_margin
-
-        # --- residuals is 1D, like: [ 11.04819 -37.72836  43.35049 -16.67031]
-        residuals = subtotal_counts - expected_counts
-
-        # --- variance is 1D, like: [12.413837 49.173948 53.980069 29.783739] ---
-        variance = (
-            opposing_margin
-            * subtotal_margin
-            * ((table_margin - opposing_margin) * (table_margin - subtotal_margin))
-            / table_margin ** 3
-        )
-
-        # --- result is scalar or 1D, depending on dimensionality of residuals ---
-        with np.errstate(divide="ignore", invalid="ignore"):
-            return residuals / np.sqrt(variance)
-
-    def _subtotal_row(self, subtotal):
-        """Return (n_cols,) ndarray of z-score `subtotal` value."""
-        # --- the weighted-counts version of this subtotal-row ---
-        subtotal_counts = np.sum(self._base_counts[subtotal.addend_idxs, :], axis=0)
-
-        # --- subtotal-margin is scalar, like 547 ---
-        subtotal_margin = np.sum(subtotal_counts)
-
-        # --- base-cols-margin is 1D because no MR dimensions ---
-        opposing_margin = self._cube_result_matrix.columns_margin
-
-        # --- table_margin is scalar because no MR dimensions ---
-        table_margin = self._table_margin
-
-        # --- expected_counts is 1D ---
-        expected_counts = opposing_margin * subtotal_margin / table_margin
-
-        # --- residuals is 1D, like: [ 11.04819 -37.72836  43.35049 -16.67031]
-        residuals = subtotal_counts - expected_counts
-
-        # --- variance is 1D, like: [12.413837 49.173948 53.980069 29.783739] ---
-        variance = (
-            opposing_margin
-            * subtotal_margin
-            * ((table_margin - opposing_margin) * (table_margin - subtotal_margin))
-            / table_margin ** 3
-        )
-
-        # --- result is scalar or 1D, depending on dimensionality of residuals ---
-        with np.errstate(divide="ignore", invalid="ignore"):
-            return residuals / np.sqrt(variance)
-
-
-# === CUBE-RESULT MATRIX OBJECTS ===
-
-
-class _BaseCubeResultMatrix(object):
+class BaseCubeResultMatrix(object):
     """Base class for all cube-result matrix (2D second-order analyzer) objects."""
 
     def __init__(self, dimensions, weighted_counts, unweighted_counts):
@@ -1051,7 +200,7 @@ class _BaseCubeResultMatrix(object):
 
     @staticmethod
     def _regular_matrix_class(dimension_types):
-        """Return _BaseCubeResultMatrix subclass appropriate to `dimension_types`."""
+        """Return BaseCubeResultMatrix subclass appropriate to `dimension_types`."""
         return (
             _MrXMrMatrix
             if dimension_types == (DT.MR, DT.MR)
@@ -1108,7 +257,7 @@ class _BaseCubeResultMatrix(object):
         return np.ix_(self._dimensions[-2].valid_elements.element_idxs)
 
 
-class _CatXCatMatrix(_BaseCubeResultMatrix):
+class _CatXCatMatrix(BaseCubeResultMatrix):
     """Matrix for CAT_X_CAT cubes and base class for most other matrix classes.
 
     Despite the name, this matrix is used for CA_SUBVAR and CA_CAT dimension too, since
